requirements.txt
streamlit>=1.28.0
qrcode[pil]>=7.4.2
plotly>=5.18.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
Pillow>=10.0.0
